+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Update unit test.
 - For `Custom` project type, some API has switched to the same as `Smart Home`.
 - Rewrite the `Custom` project start process. User will be created and authorized automatically.
+- Rewrite the `Smart Home` project start process. No need to login user anymore.
 
 ### Fixed
 - 1004 signature error when url query has more than 2 elements.